+++ sui-contracts/infinite-sea/sources/item_creation_create_logic.move
@@ -12,7 +12,7 @@ module infinite_sea::item_creation_create_logic {
         base_quantity: u32,
         base_experience: u32,
         base_creation_time: u64,
-        energy_cost: u32,
+        energy_cost: u64,
         success_rate: u16,
         item_creation_table: &item_creation::ItemCreationTable,
         ctx: &mut TxContext,
