+++ sui-contracts/infinite-sea/sources/item_production_aggregate.move
@@ -29,7 +29,7 @@ module infinite_sea::item_production_aggregate {
         base_quantity: u32,
         base_experience: u32,
         base_creation_time: u64,
-        energy_cost: u32,
+        energy_cost: u64,
         success_rate: u16,
         item_production_table: &mut item_production::ItemProductionTable,
         ctx: &mut tx_context::TxContext,
@@ -89,7 +89,7 @@ module infinite_sea::item_production_aggregate {
         base_quantity: u32,
         base_experience: u32,
         base_creation_time: u64,
-        energy_cost: u32,
+        energy_cost: u64,
         success_rate: u16,
         ctx: &mut tx_context::TxContext,
     ) {
