+++ sui-contracts/infinite-sea/sources/item_production_update_logic.move
@@ -11,7 +11,7 @@ module infinite_sea::item_production_update_logic {
         base_quantity: u32,
         base_experience: u32,
         base_creation_time: u64,
-        energy_cost: u32,
+        energy_cost: u64,
         success_rate: u16,
         item_production: &item_production::ItemProduction,
         ctx: &TxContext,
